refactor(streak): move auth to lib/common and use typed CustomAuth value
