+++ packages/pieces/community/streak/src/lib/actions/create-organization.ts
@@ -1,6 +1,6 @@
 import { HttpMethod } from '@activepieces/pieces-common';
 import { createAction, Property } from '@activepieces/pieces-framework';
-import { streakAuth } from '../../auth';
+import { streakAuth } from '../common/auth';
 import { streakApiCall } from '../common/client';
 import { teamDropdown } from '../common/props';
 
@@ -86,7 +86,7 @@ export const createOrganizationAction = createAction({
     if (description) body['description'] = description;
 
     const response = await streakApiCall<StreakOrganization>({
-      apiKey: context.auth as unknown as string,
+      apiKey: context.auth.secret_text,
       method: HttpMethod.POST,
       path: `/api/v2/teams/${teamKey}/organizations`,
       queryParams: { getIfExisting: 'true' },
