+++ packages/pieces/community/streak/src/lib/actions/create-task.ts
@@ -1,6 +1,6 @@
 import { HttpMethod } from '@activepieces/pieces-common';
 import { createAction, Property } from '@activepieces/pieces-framework';
-import { streakAuth } from '../../auth';
+import { streakAuth } from '../common/auth';
 import { streakApiCall } from '../common/client';
 import { boxDropdown, pipelineDropdown } from '../common/props';
 
@@ -55,7 +55,7 @@ export const createTaskAction = createAction({
     }
 
     const response = await streakApiCall<StreakTask>({
-      apiKey: context.auth as unknown as string,
+      apiKey: context.auth.secret_text,
       method: HttpMethod.POST,
       path: `/api/v2/boxes/${boxKey}/tasks`,
       contentType: 'application/json',
