+++ packages/pieces/community/streak/src/lib/actions/get-box.ts
@@ -1,6 +1,6 @@
 import { HttpMethod } from '@activepieces/pieces-common';
 import { createAction } from '@activepieces/pieces-framework';
-import { streakAuth } from '../../auth';
+import { streakAuth } from '../common/auth';
 import { streakApiCall } from '../common/client';
 import { flattenStreakBox } from '../common/flatten';
 import { boxDropdown, pipelineDropdown } from '../common/props';
@@ -17,7 +17,7 @@ export const getBoxAction = createAction({
   },
   async run(context) {
     const response = await streakApiCall<StreakBox>({
-      apiKey: context.auth as unknown as string,
+      apiKey: context.auth.secret_text,
       method: HttpMethod.GET,
       path: `/api/v1/boxes/${context.propsValue.boxKey}`,
     });
